fix : logic point event

+++ DemoTest/src/main/java/com/example/DemoTest/service/UserService.java
@@ -104,7 +104,7 @@ public class UserService implements UserDetailsService {
     }
 
 
-    public  Boolean updatePoint(Long user_id,Long addPointValue){
+    public  Boolean updatePoint(Long user_id,Integer addPointValue){
         User user=findUserById(user_id);
         user.setPoint(user.getPoint()+addPointValue);
         userRepository.save(user);
